fix(order-service): harden order model and test database

+++ services/order-service/src/main/java/com/diluwar/order/dto/OrderItemRequest.java
@@ -1,5 +1,7 @@
 package com.diluwar.order.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
@@ -14,6 +16,6 @@ public record OrderItemRequest(
 
         @NotNull(message = "unitPrice is required")
         @DecimalMin(value = "0.0", inclusive = false, message = "unitPrice must be greater than 0")
-        Double unitPrice
+        BigDecimal unitPrice
 ) {
 }
