+++ services/order-service/src/main/java/com/diluwar/order/dto/OrderItemResponse.java
@@ -1,9 +1,11 @@
 package com.diluwar.order.dto;
 
+import java.math.BigDecimal;
+
 public record OrderItemResponse(
         Long id,
         Long productId,
         Integer quantity,
-        Double unitPrice
+        BigDecimal unitPrice
 ) {
 }
